+++ bar/packages/desktop/src/widget_factory.rs
@@ -774,28 +774,6 @@ impl WidgetFactory {
     Ok(())
   }
 
-  /// Stops any currently running preview widget(s).
-  pub async fn stop_all_previews(&self) -> anyhow::Result<()> {
-    tracing::info!("Stopping all preview widgets.");
-
-    // Find widget states marked as previews.
-    let preview_widget_ids = {
-      let widget_states = self.widget_states.lock().await;
-      widget_states
-        .iter()
-        .filter(|(_, state)| state.is_preview)
-        .map(|(id, _)| id.clone())
-        .collect::<Vec<_>>()
-    };
-
-    // Stop each preview widget.
-    for widget_id in preview_widget_ids {
-      self.stop_by_id(&widget_id)?;
-    }
-
-    Ok(())
-  }
-
   /// Relaunches all currently open widgets.
   pub async fn relaunch_all(&self) -> anyhow::Result<()> {
     let widget_ids =